pyproject.toml
[project]
name = "media-bot"
version = "0.1.0"
description = "Smart Raspberry Pi media center with Telegram bot control"
authors = [
    {name = "Dmitri Tsiu",email = "laoqiu1015@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.12.4,<3.0.0",
    "python-telegram-bot[job-queue]>=21.0,<22.0",
    "python-mpv>=1.0.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "libtorrent>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0"
]

[tool.poetry]
packages = [
    { include = "app" }
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"


[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
